Code/Python/AiVoiceRecorder/requirements.txt
accelerate==1.10.0
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
fastapi==0.116.1
filelock==3.13.1
fsspec==2024.6.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.4
llvmlite==0.44.0
MarkupSafe==2.1.5
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.3
numba==0.61.2
numpy==2.1.2
openai-whisper==20250625
packaging==25.0
pillow==11.0.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
sniffio==1.3.1
starlette==0.47.2
sympy==1.13.1
tiktoken==0.11.0
tokenizers==0.21.4
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
tqdm==4.67.1
transformers==4.55.1
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.5.0
